# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateOnePostRequest:
    properties:
      category_id:
        example: "0"
        type: string
      content_markdown:
        type: string
      image:
        type: string
      title:
        maxLength: 225
        minLength: 1
        type: string
      visibility:
        type: boolean
    required:
    - title
    type: object
  dto.DeleteOneFileRequest:
    properties:
      bucket_name:
        maxLength: 63
        minLength: 1
        type: string
      object_name:
        maxLength: 1024
        minLength: 1
        type: string
    required:
    - bucket_name
    - object_name
    type: object
  dto.DeleteOnePostRequest:
    properties:
      id:
        example: "0"
        type: string
    required:
    - id
    type: object
  dto.GetOnePostRequest:
    properties:
      id:
        example: "0"
        type: string
    required:
    - id
    type: object
...
  dto.UpdateOnePostRequest:
    properties:
      category_id:
        example: "0"
        type: string
      content_markdown:
        type: string
      id:
        example: "0"
        type: string
      image:
        type: string
      title:
        maxLength: 255
        minLength: 0
        type: string
      visibility:
        type: boolean
    required:
    - id
    type: object
  post.PostsVO:
    description: 获取帖子时返回的响应数据
    properties:
      category_id:
        type: string
      content_html:
        description: ContentMarkdown string `json:"content_markdown"`
        type: string
      gmt_create:
        type: string
      gmt_modified:
        type: string
      id:
        type: string
      image:
        type: string
      title:
        type: string
      visibility:
        type: boolean
    type: object
  vo.Result:
    properties:
      code:
        description: 错误码
        type: integer
      data:
        description: 响应数据
      msg:
        description: 错误信息
        type: string
      requestId:
        description: 请求ID
      timeStamp:
        description: 响应时间戳
    type: object
host: localhost:9010
info:
  contact: {}
  description: This is the API documentation for Jank Blog.
  version: "1.0"
paths:
...
  /oss/deleteFile:
    post:
      consumes:
      - application/json
      description: 从 MinIO 对象存储删除文件
      parameters:
      - description: 删除文件请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteOneFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 删除文件
      tags:
      - 对象存储
  /oss/downloadFile:
    get:
      consumes:
      - application/json
      description: 从 MinIO 对象存储下载文件
      parameters:
      - in: query
        maxLength: 63
        minLength: 1
        name: bucket_name
        required: true
        type: string
      - in: query
        maxLength: 1024
        minLength: 1
        name: object_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 下载成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 文件不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 下载文件
      tags:
      - 对象存储
  /oss/listObjects:
    get:
      consumes:
      - application/json
      description: 列出 MinIO 对象存储中的文件
      parameters:
      - in: query
        maxLength: 63
        minLength: 1
        name: bucket_name
        required: true
        type: string
      - in: query
        maxLength: 1024
        name: prefix
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 列出对象
      tags:
      - 对象存储
  /oss/uploadFile:
    post:
      consumes:
      - multipart/form-data
      description: 上传文件到 MinIO 对象存储
      parameters:
      - description: 要上传的文件
        in: formData
        name: file
        required: true
        type: file
      - description: 存储桶名称
        in: formData
        name: bucket_name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 上传成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  type: string
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 上传文件
      tags:
      - 对象存储
  /post/createOnePost:
    post:
      consumes:
      - application/json
      description: 创建新的文章，支持 Markdown 格式内容，系统会自动转换为 HTML
      parameters:
      - description: 创建文章请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateOnePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/post.PostsVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 创建文章
      tags:
      - 文章
  /post/deleteOnePost:
    post:
      consumes:
      - application/json
      description: 根据文章 ID 删除指定文章
      parameters:
      - description: 删除文章请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.DeleteOnePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/vo.Result'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 删除文章
      tags:
      - 文章
  /post/getAllPosts:
    get:
      consumes:
      - application/json
      description: 获取所有的文章列表，按创建时间倒序排序
      parameters:
      - description: 页码(默认为1)
        in: query
        name: page
        type: integer
      - description: 每页条数(默认为5,最大100)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/post.PostsVO'
                  type: array
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 获取文章列表
      tags:
      - 文章
  /post/getOnePost:
    get:
      consumes:
      - application/json
      description: 根据文章 ID 获取文章的详细信息
      parameters:
      - description: 获取文章请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.GetOnePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/post.PostsVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      summary: 获取文章详情
      tags:
      - 文章
  /post/updateOnePost:
    post:
      consumes:
      - application/json
      description: 更新已存在的文章内容
      parameters:
      - description: 更新文章请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateOnePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/vo.Result'
            - properties:
                data:
                  $ref: '#/definitions/post.PostsVO'
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/vo.Result'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/vo.Result'
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/vo.Result'
      security:
      - BearerAuth: []
      summary: 更新文章
      tags:
      - 文章
...
securityDefinitions:
  BearerAuth:
    description: '输入格式: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"